+++ src/segment/SchemaSegment.hpp
@@ -11,12 +11,12 @@ public:
     /**
      * Create a schema segment with the given schema, serialize to disk.
      */
-    SchemaSegment(BufferManager& bufferManager, uint64_t segmentId, std::unique_ptr<Schema> schema);
+    SchemaSegment(BufferManager& bufferManager, uint16_t segmentId, std::unique_ptr<Schema> schema);
 
     /**
      * Deserialize schema segment from disk.
      */
-    SchemaSegment(BufferManager& bufferManager, uint64_t segmentId);
+    SchemaSegment(BufferManager& bufferManager, uint16_t segmentId);
 
     /*
      * Return reference of loaded schema.
